converts a file into ts
